+++ src/extension.ts
@@ -92,6 +92,15 @@ export function activate(context: vscode.ExtensionContext) {
             ctl.makeYamlFromImpl,
             ctl),
     );
+
+    vscode.window.registerTreeDataProvider('gnuradioModule', ctl);
+
+    context.subscriptions.push(
+        vscode.commands.registerCommand(
+            `${ctl.extId}.refreshView`,
+            ctl.refresh,
+            ctl),
+    );
 }
 
 export function deactivate() { }